lib/Makefile: set correct soname for libsign
The current soname of libsign is libsign.so:
$ readelf -d libsign.so.0.3.2 | grep SONA ME
0x000000000000000e (SONAME)    Library soname: [libsign.so]

The libsign.so is a symbolic link of libsign.so.0.3.2 and it is not
installed by default because it is packaged to dev package. Then we will
encounter an error when run command selsign:
$ selsign
selsign: error while loading shared libraries: libsign.so: cannot open shared object file: No such file or directory

$ ldd selsign |grep libsign
    libsign.so => not found

Set the soname to libsign.so.$(LIBSIGN_MAJOR_VERSION) to fix the issue.

Signed-off-by: Yi Zhao <yi.zhao@windriver.com>